Add payment success function

+++ src/app/checkout/checkout-payment/checkout-payment.component.ts
@@ -9,7 +9,13 @@ import { Cart } from '@shared/cart';
 import { SharedModule } from '@shared/shared.module';
 
 import { CheckoutService } from '../checkout.service';
-import { Stripe, StripeCardCvcElement, StripeCardExpiryElement, StripeCardNumberElement, loadStripe } from '@stripe/stripe-js';
+import {
+  Stripe,
+  StripeCardCvcElement,
+  StripeCardExpiryElement,
+  StripeCardNumberElement,
+  loadStripe,
+} from '@stripe/stripe-js';
 
 @Component({
   selector: 'sc-checkout-payment',
@@ -39,44 +45,45 @@ export class CheckoutPaymentComponent implements OnInit {
   ) {}
 
   ngOnInit(): void {
-    loadStripe('pk_test_51KfxOmGLD3kMElxsCct2QuQUasU1iM6IBqVOTPyyLbhxeWwaA6VVUdgWrrZnqlB3H9HxdQaTsaczNRw7f8Lhtyog005QXd19uy')
-      .then(stripe => {
-        this.stripe = stripe;
+    loadStripe(
+      'pk_test_51KfxOmGLD3kMElxsCct2QuQUasU1iM6IBqVOTPyyLbhxeWwaA6VVUdgWrrZnqlB3H9HxdQaTsaczNRw7f8Lhtyog005QXd19uy'
+    ).then((stripe) => {
+      this.stripe = stripe;
 
-        const elements = stripe?.elements();
+      const elements = stripe?.elements();
 
-        if (elements) {
-          this.cardNumber = elements.create('cardNumber');
-          this.cardNumber.mount(this.cardNumberElement?.nativeElement);
-          this.cardNumber.on('change', event => {
-            if (event.error) {
-              this.cardErrors = event.error.message;
-            } else {
-              this.cardErrors = null;
-            }
-          });
+      if (elements) {
+        this.cardNumber = elements.create('cardNumber');
+        this.cardNumber.mount(this.cardNumberElement?.nativeElement);
+        this.cardNumber.on('change', (event) => {
+          if (event.error) {
+            this.cardErrors = event.error.message;
+          } else {
+            this.cardErrors = null;
+          }
+        });
 
-          this.cardExpiry = elements.create('cardExpiry');
-          this.cardExpiry.mount(this.cardExpiryElement?.nativeElement);
-          this.cardExpiry.on('change', event => {
-            if (event.error) {
-              this.cardErrors = event.error.message;
-            } else {
-              this.cardErrors = null;
-            }
-          })
+        this.cardExpiry = elements.create('cardExpiry');
+        this.cardExpiry.mount(this.cardExpiryElement?.nativeElement);
+        this.cardExpiry.on('change', (event) => {
+          if (event.error) {
+            this.cardErrors = event.error.message;
+          } else {
+            this.cardErrors = null;
+          }
+        });
 
-          this.cardCvc = elements.create('cardCvc');
-          this.cardCvc.mount(this.cardCvcElement?.nativeElement);
-          this.cardCvc.on('change', event => {
-            if (event.error) {
-              this.cardErrors = event.error.message;
-            } else {
-              this.cardErrors = null;
-            }
-          })
-        }
-      })
+        this.cardCvc = elements.create('cardCvc');
+        this.cardCvc.mount(this.cardCvcElement?.nativeElement);
+        this.cardCvc.on('change', (event) => {
+          if (event.error) {
+            this.cardErrors = event.error.message;
+          } else {
+            this.cardErrors = null;
+          }
+        });
+      }
+    });
   }
 
   submitOrder() {
@@ -96,11 +103,28 @@ export class CheckoutPaymentComponent implements OnInit {
       next: (order) => {
         console.log(order);
         this._toastrService.success('Order created successfully');
-        this._cartService.deleteLocalCart();
 
-        const navigationExtras: NavigationExtras = { state: order };
+        this.stripe
+          ?.confirmCardPayment(basket.clientSecret!, {
+            payment_method: {
+              card: this.cardNumber!,
+              billing_details: {
+                name: this.checkoutForm?.get('paymentForm')?.get('nameOnCard')
+                  ?.value,
+              },
+            },
+          })
+          .then((result) => {
+            console.log(result);
 
-        this._router.navigate(['checkout/success'], navigationExtras);
+            if (result.paymentIntent?.status === 'succeeded') {
+              this._cartService.deleteLocalCart();
+
+              const navigationExtras: NavigationExtras = { state: order };
+
+              this._router.navigate(['checkout/success'], navigationExtras);
+            }
+          });
       },
       error: (error) => {
         this._toastrService.error(error);
